Вносит правки после первой проверки
