+++ src/view/film-list.js
@@ -1,12 +1,14 @@
-import {createElement} from "../utils.js";
+import {createElement} from "../dom-utils.js";
 
 const createFilmListTemplate = () => {
-  return (`<section class="films">
+  return (`
+    <section class="films">
       <section class="films-list">
         <h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>
         <div class="films-list__container"></div>
       </section>
-    </section>`);
+    </section>
+  `).trim();
 };
 
 export default class FilmList {
