+++ src/view/most-commented-list.js
@@ -1,10 +1,12 @@
-import {createElement} from "../utils.js";
+import {createElement} from "../dom-utils.js";
 
 const createMostCommentedListTemplate = () => {
-  return (`<section class="films-list films-list--extra">
+  return (`
+    <section class="films-list films-list--extra">
       <h2 class="films-list__title">Most Commented</h2>
       <div class="films-list__container"></div>
-    </section>`);
+    </section>
+  `).trim();
 };
 
 export default class MostCommentedList {
