+++ src/view/sort-filter.js
@@ -1,11 +1,13 @@
-import {createElement} from "../utils.js";
+import {createElement} from "../dom-utils.js";
 
 const createSortFilterTemplate = () => {
-  return (`<ul class="sort">
+  return (`
+    <ul class="sort">
       <li><a href="#" class="sort__button sort__button--active">Sort by default</a></li>
       <li><a href="#" class="sort__button">Sort by date</a></li>
       <li><a href="#" class="sort__button">Sort by rating</a></li>
-    </ul>`);
+    </ul>
+  `).trim();
 };
 
 export default class Sort {
