+++ src/view/user-title.js
@@ -1,10 +1,12 @@
-import {createElement} from '../utils.js';
+import {createElement} from '../dom-utils.js';
 
 const createUserTitleTemplate = () => {
-  return (`<section class="header__profile profile">
-       <p class="profile__rating">Movie Buff</p>
-       <img class="profile__avatar" src="images/bitmap@2x.png" alt="Avatar" width="35" height="35">
-      </section>`);
+  return (`
+    <section class="header__profile profile">
+      <p class="profile__rating">Movie Buff</p>
+      <img class="profile__avatar" src="images/bitmap@2x.png" alt="Avatar" width="35" height="35">
+    </section>
+  `).trim();
 };
 
 export default class UserTitle {
